+++ karma.conf.js
@@ -13,8 +13,10 @@ module.exports = (config) => {
       'node_modules/pixi.js/dist/pixi.js',
       'node_modules/react-bootstrap/dist/react-bootstrap.js',
       'node_modules/bootstrap/dist/css/bootstrap.min.css',
+      'node_modules/font-awesome/css/font-awesome.css',
       'build/hglib.css',
       'test/**/*.+(js|jsx)',
+      // 'test/AddTrackTests.js',
       // 'test/MinimalViewconfTest.js'
       // 'test/AxisTests.js',
       // 'test/PngExportTest.js',