+++ travis_build_test.sh
@@ -8,6 +8,7 @@ die() { set +v; echo "$*" 1>&2 ; sleep 1; exit 1; }
 # https://github.com/travis-ci/travis-ci/issues/6018
 
 start eslint
+
 ./node_modules/eslint/bin/eslint.js \
   karma.conf.js \
   app/scripts/configs \